Added NODE_ENV to the scripts and made it work on the bundle
Before this I was not able to differentiatea a dev build from a prod build, I added webpack DefinePlugin to mimic process.env.NODE_ENV on the bundle and choose if I want to see the redux dev tools or not.

Also added the NODE_ENV to the scripts so I can propagate it to the bundle.